requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-openai
openai
python-dotenv
pytest
fastapi
uvicorn
streamlit
python-multipart
qdrant-client
pypdf
pymupdf
redis
rank_bm25
sentence-transformers
